emr-backend/requirements.txt
aws-cdk-lib
constructs
aws-cdk.aws-appsync-alpha